Add RenderCommandBase class and related command types for rendering
